+++ src/coinbase/smart_contract.ts
@@ -5,6 +5,8 @@ import {
   SmartContractType as SmartContractTypeModel,
   SmartContractOptions as SmartContractOptionsModel,
   TokenContractOptions as TokenContractOptionsModel,
+  NFTContractOptions as NFTContractOptionsModel,
+  MultiTokenContractOptions as MultiTokenContractOptionsModel,
 } from "../client/api";
 import { Transaction } from "./transaction";
 import {
@@ -12,6 +14,7 @@ import {
   SmartContractType,
   NFTContractOptions,
   TokenContractOptions,
+  MultiTokenContractOptions,
   TransactionStatus,
 } from "./types";
 import { Coinbase } from "./coinbase";
@@ -155,6 +158,8 @@ export class SmartContract {
         return SmartContractType.ERC20;
       case SmartContractTypeModel.Erc721:
         return SmartContractType.ERC721;
+      case SmartContractTypeModel.Erc1155:
+        return SmartContractType.ERC1155;
       default:
         throw new Error(`Unknown smart contract type: ${this.model.type}`);
     }
@@ -172,12 +177,16 @@ export class SmartContract {
         symbol: this.model.options.symbol,
         totalSupply: this.model.options.total_supply,
       } as TokenContractOptions;
-    } else {
+    } else if (this.isERC721(this.getType(), this.model.options)) {
       return {
         name: this.model.options.name,
         symbol: this.model.options.symbol,
         baseURI: this.model.options.base_uri,
       } as NFTContractOptions;
+    } else {
+      return {
+        uri: this.model.options.uri,
+      } as MultiTokenContractOptions;
     }
   }
 
@@ -303,5 +312,19 @@ export class SmartContract {
     options: SmartContractOptionsModel,
   ): options is TokenContractOptionsModel {
     return type === SmartContractType.ERC20;
+  }
+
+  /**
+   * Type guard for checking if the smart contract is an ERC721.
+   *
+   * @param type - The type of the smart contract.
+   * @param options - The options of the smart contract.
+   * @returns True if the smart contract is an ERC721, false otherwise.
+   */
+  private isERC721(
+    type: SmartContractType,
+    options: SmartContractOptionsModel,
+  ): options is NFTContractOptionsModel {
+    return type === SmartContractType.ERC721;
   }
 }