+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Add `deployNFT` method to `WalletAddress` and `Wallet` to deploy an ERC721, updated `SmartContract` class to support deployment and fetching contract details
+- Add `deployMultiToken` method to `WalletAddress` and `Wallet` to deploy an ERC1155, updated `SmartContract` class to support deployment and fetching contract details
 
 ## [0.6.1] - 2024-09-23